Integración de sistema de bitácora
Se ha integrado un sistema de bitácora para registrar actividades de usuarios y su criticidad. Se añadieron las clases `Bitacora` y `EnumCriticidad` al proyecto, y se creó un diagrama de clases correspondiente. Se modificó `Usuario.cs` para cambiar el tipo de `Borrado` a `int`.

En `UsuarioService.cs`, se actualizó el método de verificación de contraseñas para registrar intentos de inicio de sesión en la bitácora. Se añadieron métodos para listar usuarios y gestionar intentos de inicio de sesión.

Se implementó `BitacoraService` para gestionar registros de bitácora, y `MP_Bitacora` para interactuar con la base de datos. Se actualizó `MP_Usuario.cs` para obtener usuarios desde la base de datos y gestionar intentos de inicio de sesión.

Se añadieron referencias a las nuevas clases en los archivos de proyecto correspondientes y se actualizó la versión de Visual Studio en la solución.

+++ BE/Usuario.cs
@@ -51,7 +51,7 @@ namespace BE
             set { salt = value; }
         }
 
-        public byte Borrado { get; set; }
+        public int Borrado { get; set; }
         
     }
 }
